Remove lint, and replace gulp with using webpack directly Switch to 4-space indenting
