--- app/templates/.eslintrc
@@ -1,11 +1,0 @@
-{
-  "parser": "babel-eslint",
-  "rules": {
-    "strict": 0,
-    "quotes": [2, "single"]
-  },
-  "env": {
-    "browser": true,
-    "node": true
-  }
-}
