--- app/templates/.jscsrc
@@ -1,6 +1,0 @@
-{
-  "preset": "google",
-  "maximumLineLength": null,
-  "esnext": true,
-  "disallowSpacesInsideObjectBrackets": null
-}
