+++ app/templates/.npmignore
@@ -1,4 +1,3 @@
-dist
 .idea
 
 # Logs
