--- app/templates/.travis.yml
@@ -1,9 +1,0 @@
-language: node_js
-node_js:
-  - "4"
-  - "stable"
-sudo: false
-script: "gulp coverage"
-after_success:
-  - npm install -g codeclimate-test-reporter
-  - codeclimate-test-reporter < coverage/lcov.info
