+++ app/templates/package.json
@@ -4,12 +4,12 @@
   "description": "<%- description %>",
   "main": "dist/<%= repo %>.js",
   "scripts": {
-    "test": "gulp",
-    "lint": "gulp lint",
-    "test-browser": "gulp test-browser",
-    "watch": "gulp watch",
-    "build": "gulp build",
-    "coverage": "gulp coverage"
+    "build": "webpack -p",
+    "build-dev": "webpack -d",
+    "test": "mocha --compilers js:babel-register test/setup/node.js test/unit/**/*.js",
+    "test:watch": "npm run test -- -w",
+    "build-browser-tests": "webpack -d --config webpack.tests.config.js",
+    "build-browser-tests:watch": "npm run build-browser-tests -- --watch"
   },
   "repository": {
     "type": "git",
@@ -23,40 +23,17 @@
   },
   "homepage": "https://github.com/<%= user %>/<%= repo %>",
   "devDependencies": {
-    "babel-core": "^6.3.26",
-    "babel-eslint": "^4.1.6",
-    "babel-loader": "^6.2.0",
+    "babel-core": "^6.5.1",
+    "babel-loader": "^6.2.2",
     "babel-polyfill": "^6.3.14",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-register": "^6.3.13",
+    "babel-preset-es2015": "^6.5.0",
+    "babel-register": "^6.5.1",
     "chai": "^3.4.1",
-    "del": "^2.2.0",
     "glob": "^6.0.3",
-    "gulp": "^3.9.0",
-    "gulp-babel": "^6.1.1",
-    "gulp-eslint": "^1.1.1",
-    "gulp-filter": "^3.0.0",
-    "gulp-istanbul": "^0.10.3",
-    "gulp-jscs": "^3.0.0",
-    "gulp-livereload": "^3.8.1",
-    "gulp-load-plugins": "^1.1.0",
-    "gulp-mocha": "^2.2.0",
-    "gulp-plumber": "^1.0.1",
-    "gulp-rename": "^1.2.2",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-uglify": "^1.5.1",
-    "gulp-util": "^3.0.7",
-    "isparta": "^4.0.0",
     "json-loader": "^0.5.3",
     "mocha": "^2.3.4",
     "sinon": "^1.17.2",
     "sinon-chai": "^2.8.0",
-    "vinyl-source-stream": "^1.1.0",
-    "webpack": "^1.12.9",
-    "webpack-stream": "^3.1.0"
-  },
-  "babelBoilerplateOptions": {
-    "entryFileName": "<%= repo %>",
-    "mainVarName": "<%= variable %>"
+    "webpack": "^1.12.9"
   }
 }
