+++ app/templates/src/index.js
@@ -1,7 +1,7 @@
 const <%= variable %> = {
-  greet() {
-    return 'hello';
-  }
+    greet() {
+        return 'hello';
+    }
 };
 
 export default <%= variable %>;
