--- app/templates/test/.eslintrc
@@ -1,14 +1,0 @@
-{
-  extends: "./setup/.globals.json",
-  "parser": "babel-eslint",
-  "rules": {
-    "strict": 0,
-    "quotes": [2, "single"],
-    "no-unused-expressions": 0
-  },
-  "env": {
-    "browser": true,
-    "node": true,
-    "mocha": true
-  }
-}
