+++ app/templates/test/setup/browser.js
@@ -2,8 +2,8 @@ var mochaGlobals = require('./.globals.json').globals;
 
 window.mocha.setup('bdd');
 window.onload = function() {
-  window.mocha.checkLeaks();
-  window.mocha.globals(Object.keys(mochaGlobals));
-  window.mocha.run();
-  require('./setup')(window);
+    window.mocha.checkLeaks();
+    window.mocha.globals(Object.keys(mochaGlobals));
+    window.mocha.run();
+    require('./setup')(window);
 };
