+++ app/templates/test/unit/index.js
@@ -1,18 +1,18 @@
 import <%= variable %> from '../../src/<%= repo %>';
 
 describe('<%= variable %>', () => {
-  describe('Greet function', () => {
-    beforeEach(() => {
-      spy(<%= variable %>, 'greet');
-      <%= variable %>.greet();
-    });
+    describe('Greet function', () => {
+        beforeEach(() => {
+            spy(<%= variable %>, 'greet');
+            <%= variable %>.greet();
+        });
 
-    it('should have been run once', () => {
-      expect(<%= variable %>.greet).to.have.been.calledOnce;
-    });
+        it('should have been run once', () => {
+            expect(<%= variable %>.greet).to.have.been.calledOnce;
+        });
 
-    it('should have always returned hello', () => {
-      expect(<%= variable %>.greet).to.have.always.returned('hello');
+        it('should have always returned hello', () => {
+            expect(<%= variable %>.greet).to.have.always.returned('hello');
+        });
     });
-  });
 });
